Created font licensing information page
As I'd mentioned in 47ac62f, I think it's important that we publish
information on the licenses of the various fonts used.

In the case of Creative Commons Attributions licenses, this is expressly
required: "You must give appropriate credit, provide a link to the
license" (from https://creativecommons.org/licenses/by/3.0/).

I've created a page that details the fonts used, and the licenses
they're released under.  We should probably link to this page from the
copyright page of the site.